macos/requirements.txt
"flask>=2.0.0
flask-cors>=3.0.0
flask-mail>=0.9.0
requests>=2.25.0
pymysql>=1.0.0
python-dotenv>=0.19.0
boto3>=1.20.0
botocore>=1.23.0
openai>=0.25.0
Pillow>=9.0.0
mss>=7.0.0
cryptography>=37.0.0
pyautogui>=0.9.0
psutil>=5.9.0
pywebview>=4.0.0
rumps>=0.4.0
pyobjc-framework-Quartz>=9.0
pyobjc-framework-AppKit>=9.0
pyinstaller>=5.0
"